server/requirements.txt
anyio==3.7.0
boto3==1.26.165
botocore==1.29.165
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
exceptiongroup==1.1.2
fastapi==0.99.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
psycopg2-binary==2.9.6
pydantic==1.10.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
